+++ packages/tailwindcss/src/intellisense.test.ts
@@ -3,22 +3,18 @@ import { buildDesignSystem } from './design-system'
 import { Theme } from './theme'
 
 function loadDesignSystem() {
-  return buildDesignSystem(
-    new Theme(
-      new Map([
-        ['--spacing-0_5', '0.125rem'],
-        ['--spacing-1', '0.25rem'],
-        ['--spacing-3', '0.75rem'],
-        ['--spacing-4', '1rem'],
-        ['--width-4', '1rem'],
-        ['--colors-red-500', 'red'],
-        ['--colors-blue-500', 'blue'],
-        ['--breakpoint-sm', '640px'],
-        ['--font-size-xs', '0.75rem'],
-        ['--font-size-xs--line-height', '1rem'],
-      ]),
-    ),
-  )
+  let theme = new Theme()
+  theme.add('--spacing-0_5', '0.125rem')
+  theme.add('--spacing-1', '0.25rem')
+  theme.add('--spacing-3', '0.75rem')
+  theme.add('--spacing-4', '1rem')
+  theme.add('--width-4', '1rem')
+  theme.add('--colors-red-500', 'red')
+  theme.add('--colors-blue-500', 'blue')
+  theme.add('--breakpoint-sm', '640px')
+  theme.add('--font-size-xs', '0.75rem')
+  theme.add('--font-size-xs--line-height', '1rem')
+  return buildDesignSystem(theme)
 }
 
 test('getClassList', () => {